break out StepSerializer
